Removed default window from within graphic_context.
The window was necessary only for the creation, but a temporary window
works just fine for that.
Having a window object inside the context caused issue with SDL events,
since the QUIT event is generated when the last window is destroyed.
The presence of the dummy window inside the graphic_context prevented
the generation of QUIT events.